Updating LapackWrappers to start supporting complex (not tested yet)

+++ TPLs/LapackWrappers/LapackWrappers.float.cpp
@@ -86,9 +86,9 @@ void Lapack<float>::copy( int N, const float *DX, int INCX, float *DY, int INCY 
     cblas_scopy( N, DX, INCX, DY, INCY );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Nl = N, INCXl = INCX, INCYl = INCY;
-    FORTRAN_WRAPPER(::scopy )( &Nl, (float *) DX, &INCXl, DY, &INCYl );
+    FORTRAN_WRAPPER( ::scopy )( &Nl, (float *) DX, &INCXl, DY, &INCYl );
 #else
-    FORTRAN_WRAPPER(::scopy )( &N, (float *) DX, &INCX, DY, &INCY );
+    FORTRAN_WRAPPER( ::scopy )( &N, (float *) DX, &INCX, DY, &INCY );
 #endif
 }
 // Define the member functions
@@ -104,9 +104,9 @@ void Lapack<float>::swap( int N, float *DX, int INCX, float *DY, int INCY )
     cblas_sswap( N, DX, INCX, DY, INCY );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Nl = N, INCXl = INCX, INCYl = INCY;
-    FORTRAN_WRAPPER(::sswap )( &Nl, DX, &INCXl, DY, &INCYl );
+    FORTRAN_WRAPPER( ::sswap )( &Nl, DX, &INCXl, DY, &INCYl );
 #else
-    FORTRAN_WRAPPER(::sswap )( &N, DX, &INCX, DY, &INCY );
+    FORTRAN_WRAPPER( ::sswap )( &N, DX, &INCX, DY, &INCY );
 #endif
 }
 #undef sscal
@@ -121,14 +121,14 @@ void Lapack<float>::scal( int N, float DA, float *DX, int INCX )
     cblas_sscal( N, DA, DX, INCX );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, INCXp = INCX;
-    FORTRAN_WRAPPER(::sscal )( &Np, &DA, DX, &INCXp );
+    FORTRAN_WRAPPER( ::sscal )( &Np, &DA, DX, &INCXp );
 #else
-    FORTRAN_WRAPPER(::sscal )( &N, &DA, DX, &INCX );
+    FORTRAN_WRAPPER( ::sscal )( &N, &DA, DX, &INCX );
 #endif
 }
 #undef snrm2
 template<>
-float Lapack<float>::nrm2( int N, const float *DX, int INCX )
+double Lapack<float>::nrm2( int N, const float *DX, int INCX )
 {
 #ifdef USE_ATLAS
     return cblas_snrm2( N, DX, INCX );
@@ -138,9 +138,9 @@ float Lapack<float>::nrm2( int N, const float *DX, int INCX )
     return cblas_snrm2( N, DX, INCX );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, INCXp = INCX;
-    return FORTRAN_WRAPPER(::snrm2 )( &Np, (float *) DX, &INCXp );
+    return FORTRAN_WRAPPER( ::snrm2 )( &Np, (float *) DX, &INCXp );
 #else
-    return FORTRAN_WRAPPER(::snrm2 )( &N, (float *) DX, &INCX );
+    return FORTRAN_WRAPPER( ::snrm2 )( &N, (float *) DX, &INCX );
 #endif
 }
 #undef isamax
@@ -155,9 +155,9 @@ int Lapack<float>::iamax( int N, const float *DX, int INCX )
     return cblas_isamax( N, DX, INCX ) - 1;
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, INCXp = INCX;
-    return FORTRAN_WRAPPER(::isamax )( &Np, (float *) DX, &INCXp ) - 1;
+    return FORTRAN_WRAPPER( ::isamax )( &Np, (float *) DX, &INCXp ) - 1;
 #else
-    return FORTRAN_WRAPPER(::isamax )( &N, (float *) DX, &INCX ) - 1;
+    return FORTRAN_WRAPPER( ::isamax )( &N, (float *) DX, &INCX ) - 1;
 #endif
 }
 #undef saxpy
@@ -172,9 +172,9 @@ void Lapack<float>::axpy( int N, float DA, const float *DX, int INCX, float *DY,
     cblas_saxpy( N, DA, DX, INCX, DY, INCY );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, INCXp = INCX, INCYp = INCY;
-    FORTRAN_WRAPPER(::saxpy )( &Np, &DA, (float *) DX, &INCXp, DY, &INCYp );
+    FORTRAN_WRAPPER( ::saxpy )( &Np, &DA, (float *) DX, &INCXp, DY, &INCYp );
 #else
-    FORTRAN_WRAPPER(::saxpy )( &N, &DA, (float *) DX, &INCX, DY, &INCY );
+    FORTRAN_WRAPPER( ::saxpy )( &N, &DA, (float *) DX, &INCX, DY, &INCY );
 #endif
 }
 #undef sgemv
@@ -194,10 +194,10 @@ void Lapack<float>::gemv( char TRANS, int M, int N, float ALPHA, const float *A,
     cblas_sgemv( CblasColMajor, TRANS2( TRANS ), M, N, ALPHA, A, LDA, DX, INCX, BETA, DY, INCY );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Mp = M, Np = N, LDAp = LDA, INCXp = INCX, INCYp = INCY;
-    FORTRAN_WRAPPER(::sgemv )
+    FORTRAN_WRAPPER( ::sgemv )
     ( &TRANS, &Mp, &Np, &ALPHA, (float *) A, &LDAp, (float *) DX, &INCXp, &BETA, DY, &INCYp );
 #else
-    FORTRAN_WRAPPER(::sgemv )
+    FORTRAN_WRAPPER( ::sgemv )
     ( &TRANS, &M, &N, &ALPHA, (float *) A, &LDA, (float *) DX, &INCX, &BETA, DY, &INCY );
 #endif
 }
@@ -210,7 +210,7 @@ void Lapack<float>::gemm( char TRANSA, char TRANSB, int M, int N, int K, float A
     cblas_sgemm( CblasColMajor, (CBLAS_TRANSPOSE) TRANSA, (CBLAS_TRANSPOSE) TRANSB, M, N, K, ALPHA,
         A, LDA, B, LDB, BETA, C, LDC );
 #elif defined( USE_ACML )
-    FORTRAN_WRAPPER(::sgemm )
+    FORTRAN_WRAPPER( ::sgemm )
     ( &TRANSA, &TRANSB, &M, &N, &K, &ALPHA, (float *) A, &LDA, (float *) B, &LDB, &BETA, C, &LDC, 1,
         1 );
 //::sgemm(TRANSA,TRANSA,M,N,K,ALPHA,(float*)A,LDA,(float*)B,LDB,BETA,C,LDC);
@@ -222,17 +222,17 @@ void Lapack<float>::gemm( char TRANSA, char TRANSB, int M, int N, int K, float A
         BETA, C, LDC );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Mp = M, Np = N, Kp = K, LDAp = LDA, LDBp = LDB, LDCp = LDC;
-    FORTRAN_WRAPPER(::sgemm )
+    FORTRAN_WRAPPER( ::sgemm )
     ( &TRANSA, &TRANSB, &Mp, &Np, &Kp, &ALPHA, (float *) A, &LDAp, (float *) B, &LDBp, &BETA, C,
         &LDCp );
 #else
-    FORTRAN_WRAPPER(::sgemm )
+    FORTRAN_WRAPPER( ::sgemm )
     ( &TRANSA, &TRANSB, &M, &N, &K, &ALPHA, (float *) A, &LDA, (float *) B, &LDB, &BETA, C, &LDC );
 #endif
 }
 #undef sasum
 template<>
-float Lapack<float>::asum( int N, const float *DX, int INCX )
+double Lapack<float>::asum( int N, const float *DX, int INCX )
 {
 #ifdef USE_ATLAS
     return cblas_sasum( N, DX, INCX );
@@ -242,9 +242,9 @@ float Lapack<float>::asum( int N, const float *DX, int INCX )
     return cblas_sasum( N, DX, INCX );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, INCXp = INCX;
-    return FORTRAN_WRAPPER(::sasum )( &Np, (float *) DX, &INCXp );
+    return FORTRAN_WRAPPER( ::sasum )( &Np, (float *) DX, &INCXp );
 #else
-    return FORTRAN_WRAPPER(::sasum )( &N, (float *) DX, &INCX );
+    return FORTRAN_WRAPPER( ::sasum )( &N, (float *) DX, &INCX );
 #endif
 }
 #undef sdot
@@ -259,9 +259,9 @@ float Lapack<float>::dot( int N, const float *DX, int INCX, const float *DY, int
     return cblas_sdot( N, DX, INCX, DY, INCY );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, INCXp = INCX, INCYp = INCY;
-    return FORTRAN_WRAPPER(::sdot )( &Np, (float *) DX, &INCXp, (float *) DY, &INCYp );
+    return FORTRAN_WRAPPER( ::sdot )( &Np, (float *) DX, &INCXp, (float *) DY, &INCYp );
 #else
-    return FORTRAN_WRAPPER(::sdot )( &N, (float *) DX, &INCX, (float *) DY, &INCY );
+    return FORTRAN_WRAPPER( ::sdot )( &N, (float *) DX, &INCX, (float *) DY, &INCY );
 #endif
 }
 #undef sger
@@ -277,10 +277,10 @@ void Lapack<float>::ger( int N, int M, float alpha, const float *x, int INCX, co
     cblas_sger( CblasColMajor, N, M, alpha, x, INCX, y, INCY, A, LDA );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, Mp = M, INCXp = INCX, INCYp = INCY, LDAp = LDA;
-    FORTRAN_WRAPPER(::sger )
+    FORTRAN_WRAPPER( ::sger )
     ( &Np, &Mp, &alpha, (float *) x, &INCXp, (float *) y, &INCYp, A, &LDAp );
 #else
-    FORTRAN_WRAPPER(::sger )( &N, &M, &alpha, (float *) x, &INCX, (float *) y, &INCY, A, &LDA );
+    FORTRAN_WRAPPER( ::sger )( &N, &M, &alpha, (float *) x, &INCX, (float *) y, &INCY, A, &LDA );
 #endif
 }
 #undef sgesv
@@ -297,14 +297,14 @@ void Lapack<float>::gesv(
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, NRHSp = NRHS, LDAp = LDA, LDBp = LDB, INFOp;
     ptrdiff_t *IPIVp = new ptrdiff_t[N];
-    FORTRAN_WRAPPER(::sgesv )( &Np, &NRHSp, A, &LDAp, IPIVp, B, &LDBp, &INFOp );
+    FORTRAN_WRAPPER( ::sgesv )( &Np, &NRHSp, A, &LDAp, IPIVp, B, &LDBp, &INFOp );
     for ( int i = 0; i < N; i++ ) {
         IPIV[i] = static_cast<int>( IPIVp[i] );
     }
     delete[] IPIVp;
     INFO         = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgesv )( &N, &NRHS, (float *) A, &LDA, IPIV, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgesv )( &N, &NRHS, (float *) A, &LDA, IPIV, B, &LDB, &INFO );
 #endif
 }
 #undef sgtsv
@@ -315,15 +315,15 @@ void Lapack<float>::gtsv(
 #ifdef USE_ATLAS
     throw std::logic_error( "ATLAS does not support sgtsv" );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgtsv )( &N, &NRHS, DL, D, DU, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgtsv )( &N, &NRHS, DL, D, DU, B, &LDB, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgtsv )( &N, &NRHS, DL, D, DU, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgtsv )( &N, &NRHS, DL, D, DU, B, &LDB, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t N2 = N, NRHS2 = NRHS, LDB2 = LDB, INFOp;
-    FORTRAN_WRAPPER(::sgtsv )( &N2, &NRHS2, DL, D, DU, B, &LDB2, &INFOp );
+    FORTRAN_WRAPPER( ::sgtsv )( &N2, &NRHS2, DL, D, DU, B, &LDB2, &INFOp );
     INFO         = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgtsv )( &N, &NRHS, DL, D, DU, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgtsv )( &N, &NRHS, DL, D, DU, B, &LDB, &INFO );
 #endif
 }
 #undef sgbsv
@@ -334,20 +334,20 @@ void Lapack<float>::gbsv(
 #ifdef USE_ATLAS
     throw std::logic_error( "ATLAS does not support sgbsv" );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgbsv )( &N, &KL, &KU, &NRHS, AB, &LDAB, IPIV, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgbsv )( &N, &KL, &KU, &NRHS, AB, &LDAB, IPIV, B, &LDB, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgbsv )( &N, &KL, &KU, &NRHS, AB, &LDAB, IPIV, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgbsv )( &N, &KL, &KU, &NRHS, AB, &LDAB, IPIV, B, &LDB, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, KLp = KL, KUp = KU, NRHSp = NRHS, LDABp = LDAB, LDBp = LDB, INFOp;
     ptrdiff_t *IPIVp = new ptrdiff_t[N];
-    FORTRAN_WRAPPER(::sgbsv )( &Np, &KLp, &KUp, &NRHSp, AB, &LDABp, IPIVp, B, &LDBp, &INFOp );
+    FORTRAN_WRAPPER( ::sgbsv )( &Np, &KLp, &KUp, &NRHSp, AB, &LDABp, IPIVp, B, &LDBp, &INFOp );
     for ( int i = 0; i < N; i++ ) {
         IPIV[i] = static_cast<int>( IPIVp[i] );
     }
     delete[] IPIVp;
     INFO         = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgbsv )( &N, &KL, &KU, &NRHS, AB, &LDAB, IPIV, B, &LDB, &INFO );
+    FORTRAN_WRAPPER( ::sgbsv )( &N, &KL, &KU, &NRHS, AB, &LDAB, IPIV, B, &LDB, &INFO );
 #endif
 }
 #undef sgetrf
@@ -357,20 +357,20 @@ void Lapack<float>::getrf( int M, int N, float *A, int LDA, int *IPIV, int &INFO
 #ifdef USE_ATLAS
     INFO = clapack_sgetrf( CblasColMajor, M, N, A, LDA, IPIV );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgetrf )( &M, &N, A, &LDA, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgetrf )( &M, &N, A, &LDA, IPIV, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgetrf )( &M, &N, A, &LDA, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgetrf )( &M, &N, A, &LDA, IPIV, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, Mp = M, LDAp = LDA, INFOp;
     ptrdiff_t *IPIVp = new ptrdiff_t[N];
-    FORTRAN_WRAPPER(::sgetrf )( &Mp, &Np, A, &LDAp, IPIVp, &INFOp );
+    FORTRAN_WRAPPER( ::sgetrf )( &Mp, &Np, A, &LDAp, IPIVp, &INFOp );
     for ( int i = 0; i < N; i++ ) {
         IPIV[i] = static_cast<int>( IPIVp[i] );
     }
     delete[] IPIVp;
     INFO             = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgetrf )( &M, &N, A, &LDA, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgetrf )( &M, &N, A, &LDA, IPIV, &INFO );
 #endif
 }
 #undef sgttrf
@@ -380,20 +380,20 @@ void Lapack<float>::gttrf( int N, float *DL, float *D, float *DU, float *DU2, in
 #ifdef USE_ATLAS
     throw std::logic_error( "ATLAS does not support sgttrf" );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgttrf )( &N, DL, D, DU, DU2, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgttrf )( &N, DL, D, DU, DU2, IPIV, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgttrf )( &N, DL, D, DU, DU2, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgttrf )( &N, DL, D, DU, DU2, IPIV, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np     = N, INFOp;
     ptrdiff_t *IPIVp = new ptrdiff_t[N];
-    FORTRAN_WRAPPER(::sgttrf )( &Np, DL, D, DU, DU2, IPIVp, &INFOp );
+    FORTRAN_WRAPPER( ::sgttrf )( &Np, DL, D, DU, DU2, IPIVp, &INFOp );
     for ( int i = 0; i < N; i++ ) {
         IPIV[i] = static_cast<int>( IPIVp[i] );
     }
     delete[] IPIVp;
     INFO         = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgttrf )( &N, DL, D, DU, DU2, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgttrf )( &N, DL, D, DU, DU2, IPIV, &INFO );
 #endif
 }
 #undef sgbtrf
@@ -403,13 +403,13 @@ void Lapack<float>::gbtrf( int M, int N, int KL, int KU, float *AB, int LDAB, in
 #ifdef USE_ATLAS
     throw std::logic_error( "ATLAS does not support sgbtrf" );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Mp = M, Np = N, KLp = KL, KUp = KU, LDABp = LDAB, INFOp;
     ptrdiff_t *IPIVp = new ptrdiff_t[N];
-    FORTRAN_WRAPPER(::sgbtrf )( &Mp, &Np, &KLp, &KUp, AB, &LDABp, IPIVp, &INFOp );
+    FORTRAN_WRAPPER( ::sgbtrf )( &Mp, &Np, &KLp, &KUp, AB, &LDABp, IPIVp, &INFOp );
     for ( int i = 0; i < N; i++ ) {
         IPIV[i] = static_cast<int>( IPIVp[i] );
     }
@@ -417,10 +417,10 @@ void Lapack<float>::gbtrf( int M, int N, int KL, int KU, float *AB, int LDAB, in
     INFO = static_cast<int>( INFOp );
 #elif defined( USE_ACML )
     d_mutex.lock();
-    FORTRAN_WRAPPER(::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
     d_mutex.unlock();
 #else
-    FORTRAN_WRAPPER(::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
+    FORTRAN_WRAPPER( ::sgbtrf )( &M, &N, &KL, &KU, AB, &LDAB, IPIV, &INFO );
 #endif
 }
 #undef sgetrs
@@ -433,10 +433,10 @@ void Lapack<float>::getrs( char TRANS, int N, int NRHS, const float *A, int LDA,
 #elif defined( USE_ACML )
     ::sgetrs( TRANS, N, NRHS, (float *) A, LDA, (int *) IPIV, B, LDB, &INFO );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgetrs )
+    FORTRAN_WRAPPER( ::sgetrs )
     ( &TRANS, &N, &NRHS, (float *) A, &LDA, (int *) IPIV, B, &LDB, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgetrs )
+    FORTRAN_WRAPPER( ::sgetrs )
     ( &TRANS, &N, &NRHS, (float *) A, &LDA, (int *) IPIV, B, &LDB, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, NRHSp = NRHS, LDAp = LDA, LDBp = LDB, INFOp;
@@ -444,11 +444,11 @@ void Lapack<float>::getrs( char TRANS, int N, int NRHS, const float *A, int LDA,
     for ( int i = 0; i < N; i++ ) {
         IPIVp[i] = IPIV[i];
     }
-    FORTRAN_WRAPPER(::sgetrs )( &TRANS, &Np, &NRHSp, (float *) A, &LDAp, IPIVp, B, &LDBp, &INFOp );
+    FORTRAN_WRAPPER( ::sgetrs )( &TRANS, &Np, &NRHSp, (float *) A, &LDAp, IPIVp, B, &LDBp, &INFOp );
     delete[] IPIVp;
     INFO         = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgetrs )
+    FORTRAN_WRAPPER( ::sgetrs )
     ( &TRANS, &N, &NRHS, (float *) A, &LDA, (int *) IPIV, B, &LDB, &INFO );
 #endif
 }
@@ -463,11 +463,11 @@ void Lapack<float>::gttrs( char TRANS, int N, int NRHS, const float *DL, const f
     ::sgttrs( TRANS, N, NRHS, (float *) DL, (float *) D, (float *) DU, (float *) DU2, (int *) IPIV,
         B, LDB, &INFO );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgttrs )
+    FORTRAN_WRAPPER( ::sgttrs )
     ( &TRANS, &N, &NRHS, (float *) DL, (float *) D, (float *) DU, (float *) DU2, (int *) IPIV, B,
         &LDB, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgttrs )
+    FORTRAN_WRAPPER( ::sgttrs )
     ( &TRANS, &N, &NRHS, (float *) DL, (float *) D, (float *) DU, (float *) DU2, (int *) IPIV, B,
         &LDB, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
@@ -476,13 +476,13 @@ void Lapack<float>::gttrs( char TRANS, int N, int NRHS, const float *DL, const f
     for ( int i = 0; i < N; i++ ) {
         IPIVp[i] = IPIV[i];
     }
-    FORTRAN_WRAPPER(::sgttrs )
+    FORTRAN_WRAPPER( ::sgttrs )
     ( &TRANS, &Np, &NRHSp, (float *) DL, (float *) D, (float *) DU, (float *) DU2, IPIVp, B, &LDBp,
         &INFOp );
     delete[] IPIVp;
     INFO         = static_cast<int>( INFOp );
 #else
-    FORTRAN_WRAPPER(::sgttrs )
+    FORTRAN_WRAPPER( ::sgttrs )
     ( &TRANS, &N, &NRHS, (float *) DL, (float *) D, (float *) DU, (float *) DU2, (int *) IPIV, B,
         &LDB, &INFO );
 #endif
@@ -497,10 +497,10 @@ void Lapack<float>::gbtrs( char TRANS, int N, int KL, int KU, int NRHS, const fl
 #elif defined( USE_ACML )
     ::sgbtrs( TRANS, N, KL, KU, NRHS, (float *) AB, LDAB, (int *) IPIV, B, LDB, &INFO );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgbtrs )
+    FORTRAN_WRAPPER( ::sgbtrs )
     ( &TRANS, &N, &KL, &KU, &NRHS, (float *) AB, &LDAB, (int *) IPIV, B, &LDB, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgbtrs )
+    FORTRAN_WRAPPER( ::sgbtrs )
     ( &TRANS, &N, &KL, &KU, &NRHS, (float *) AB, &LDAB, (int *) IPIV, B, &LDB, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, KLp = KL, KUp = KU, NRHSp = NRHS, LDABp = LDAB, LDBp = LDB, INFOp;
@@ -508,12 +508,12 @@ void Lapack<float>::gbtrs( char TRANS, int N, int KL, int KU, int NRHS, const fl
     for ( int i = 0; i < N; i++ ) {
         IPIVp[i] = IPIV[i];
     }
-    FORTRAN_WRAPPER(::sgbtrs )
+    FORTRAN_WRAPPER( ::sgbtrs )
     ( &TRANS, &Np, &KLp, &KUp, &NRHSp, (float *) AB, &LDABp, IPIVp, B, &LDBp, &INFOp );
     INFO = static_cast<int>( INFOp );
     delete[] IPIVp;
 #else
-    FORTRAN_WRAPPER(::sgbtrs )
+    FORTRAN_WRAPPER( ::sgbtrs )
     ( &TRANS, &N, &KL, &KU, &NRHS, (float *) AB, &LDAB, (int *) IPIV, B, &LDB, &INFO );
 #endif
 }
@@ -527,20 +527,20 @@ void Lapack<float>::getri(
 #elif defined( USE_ACML )
     ::sgetri_( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
 #elif defined( USE_VECLIB )
-    FORTRAN_WRAPPER(::sgetri )( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
+    FORTRAN_WRAPPER( ::sgetri )( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
 #elif defined( USE_OPENBLAS )
-    FORTRAN_WRAPPER(::sgetri )( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
+    FORTRAN_WRAPPER( ::sgetri )( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Np = N, LDAp = LDA, LWORKp = LWORK, INFOp;
     ptrdiff_t *IPIVp = new ptrdiff_t[N];
     for ( int i = 0; i < N; i++ ) {
         IPIVp[i] = IPIV[i];
     }
-    FORTRAN_WRAPPER(::sgetri )( &Np, A, &LDAp, IPIVp, WORK, &LWORKp, &INFOp );
+    FORTRAN_WRAPPER( ::sgetri )( &Np, A, &LDAp, IPIVp, WORK, &LWORKp, &INFOp );
     INFO = static_cast<int>( INFOp );
     delete[] IPIVp;
 #else
-    FORTRAN_WRAPPER(::sgetri )( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
+    FORTRAN_WRAPPER( ::sgetri )( &N, A, &LDA, (int *) IPIV, WORK, &LWORK, &INFO );
 #endif
 }
 #undef strsm
@@ -562,10 +562,10 @@ void Lapack<float>::trsm( char SIDE, char UPLO, char TRANS, char DIAG, int M, in
         ALPHA, (float *) A, LDA, B, LDB );
 #elif defined( USE_MATLAB_LAPACK )
     ptrdiff_t Mp = M, Np = N, LDAp = LDA, LDBp = LDB;
-    FORTRAN_WRAPPER(::strsm )
+    FORTRAN_WRAPPER( ::strsm )
     ( &SIDE, &UPLO, &TRANS, &DIAG, &Mp, &Np, &ALPHA, (float *) A, &LDAp, B, &LDBp );
 #else
-    FORTRAN_WRAPPER(::strsm )
+    FORTRAN_WRAPPER( ::strsm )
     ( &SIDE, &UPLO, &TRANS, &DIAG, &M, &N, &ALPHA, (float *) A, &LDA, B, &LDB );
 #endif
 }
